Refactor: move exceptions and dependencies to core
