+++ src/InodeCache.cpp
@@ -18,14 +18,14 @@
 
 #include "InodeCache.hpp"
 
-#include "Config.hpp"
-#include "core/Fd.hpp"
 #include "Finalizer.hpp"
 #include "Hash.hpp"
-#include "Logging.hpp"
-#include "Stat.hpp"
-#include "TemporaryFile.hpp"
 #include "Util.hpp"
+#include "core/Config.hpp"
+#include "core/Fd.hpp"
+#include "core/Logging.hpp"
+#include "core/Stat.hpp"
+#include "core/TemporaryFile.hpp"
 #include "core/fmtmacros.hpp"
 
 #include <atomic>
