+++ src/Manifest.cpp
@@ -18,16 +18,16 @@
 
 #include "Manifest.hpp"
 
-#include "AtomicFile.hpp"
+#include "core/AtomicFile.hpp"
 #include "CacheEntryReader.hpp"
 #include "CacheEntryWriter.hpp"
 #include "Checksum.hpp"
-#include "Config.hpp"
+#include "core/Config.hpp"
 #include "Context.hpp"
 #include "Digest.hpp"
 #include "core/File.hpp"
 #include "Hash.hpp"
-#include "Logging.hpp"
+#include "core/Logging.hpp"
 #include "core/Sloppy.hpp"
 #include "core/StdMakeUnique.hpp"
 #include "core/fmtmacros.hpp"
