+++ src/NullDecompressor.cpp
@@ -18,7 +18,7 @@
 
 #include "NullDecompressor.hpp"
 
-#include "exceptions.hpp"
+#include "core/exceptions.hpp"
 
 NullDecompressor::NullDecompressor(FILE* stream) : m_stream(stream)
 {
