+++ src/Result.cpp
@@ -18,19 +18,19 @@
 
 #include "Result.hpp"
 
-#include "AtomicFile.hpp"
 #include "CacheEntryReader.hpp"
 #include "CacheEntryWriter.hpp"
-#include "Config.hpp"
 #include "Context.hpp"
+#include "Util.hpp"
+#include "core/AtomicFile.hpp"
+#include "core/Config.hpp"
 #include "core/Fd.hpp"
 #include "core/File.hpp"
-#include "Logging.hpp"
-#include "Stat.hpp"
-#include "Statistics.hpp"
-#include "Util.hpp"
+#include "core/Logging.hpp"
+#include "core/Stat.hpp"
+#include "core/Statistics.hpp"
+#include "core/exceptions.hpp"
 #include "core/fmtmacros.hpp"
-#include "exceptions.hpp"
 
 #include <algorithm>
