+++ src/ccache.cpp
@@ -29,8 +29,8 @@
 #include "core/File.hpp"
 #include "Finalizer.hpp"
 #include "Hash.hpp"
-#include "Lockfile.hpp"
-#include "Logging.hpp"
+#include "core/Lockfile.hpp"
+#include "core/Logging.hpp"
 #include "Manifest.hpp"
 #include "MiniTrace.hpp"
 #include "ProgressBar.hpp"
@@ -39,7 +39,7 @@
 #include "ResultExtractor.hpp"
 #include "ResultRetriever.hpp"
 #include "SignalHandler.hpp"
-#include "TemporaryFile.hpp"
+#include "core/TemporaryFile.hpp"
 #include "UmaskScope.hpp"
 #include "Util.hpp"
 #include "argprocessing.hpp"
@@ -50,7 +50,7 @@
 #include "core/Sloppy.hpp"
 #include "core/StdMakeUnique.hpp"
 #include "core/fmtmacros.hpp"
-#include "exceptions.hpp"
+#include "core/exceptions.hpp"
 #include "execute.hpp"
 #include "hashutil.hpp"
 #include "language.hpp"
