+++ src/core/include/core/AtomicFile.hpp
