+++ src/core/include/core/Statistics.hpp
@@ -20,7 +20,7 @@
 
 #include "core/system.hpp"
 
-#include "Counters.hpp"
+#include "core/Counters.hpp"
 
 #include "third_party/nonstd/optional.hpp"
