+++ src/execute.cpp
@@ -19,14 +19,14 @@
 
 #include "execute.hpp"
 
-#include "Config.hpp"
 #include "Context.hpp"
-#include "core/Fd.hpp"
-#include "Logging.hpp"
 #include "SignalHandler.hpp"
-#include "Stat.hpp"
-#include "TemporaryFile.hpp"
 #include "Util.hpp"
+#include "core/Config.hpp"
+#include "core/Fd.hpp"
+#include "core/Logging.hpp"
+#include "core/Stat.hpp"
+#include "core/TemporaryFile.hpp"
 #include "core/fmtmacros.hpp"
 
 #ifdef _WIN32
