+++ src/hashutil.cpp
@@ -19,12 +19,12 @@
 #include "hashutil.hpp"
 
 #include "Args.hpp"
-#include "Config.hpp"
 #include "Context.hpp"
 #include "Hash.hpp"
-#include "Logging.hpp"
-#include "Stat.hpp"
+#include "core/Config.hpp"
+#include "core/Logging.hpp"
 #include "core/Sloppy.hpp"
+#include "core/Stat.hpp"
 #include "core/fmtmacros.hpp"
 #include "execute.hpp"
 #include "macroskip.hpp"
