+++ unittest/main.cpp
@@ -16,8 +16,8 @@
 // this program; if not, write to the Free Software Foundation, Inc., 51
 // Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
 
-#include "../src/Util.hpp"
 #include "TestUtil.hpp"
+#include "core/Util.hpp"
 #include "core/fmtmacros.hpp"
 
 #include "third_party/fmt/core.h"
