+++ unittest/test_Config.cpp
@@ -16,11 +16,11 @@
 // this program; if not, write to the Free Software Foundation, Inc., 51
 // Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
 
-#include "../src/Config.hpp"
-#include "../src/Util.hpp"
-#include "../src/exceptions.hpp"
 #include "TestUtil.hpp"
+#include "core/Config.hpp"
 #include "core/Sloppy.hpp"
+#include "core/Util.hpp"
+#include "core/exceptions.hpp"
 #include "core/fmtmacros.hpp"
 
 #include "third_party/doctest.h"
