+++ unittest/test_Stat.cpp
@@ -16,9 +16,9 @@
 // this program; if not, write to the Free Software Foundation, Inc., 51
 // Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
 
-#include "../src/Stat.hpp"
-#include "../src/Util.hpp"
 #include "TestUtil.hpp"
+#include "core/Stat.hpp"
+#include "core/Util.hpp"
 
 #include "third_party/doctest.h"
