+++ unittest/test_Statistics.cpp
@@ -16,9 +16,9 @@
 // this program; if not, write to the Free Software Foundation, Inc., 51
 // Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA
 
-#include "../src/Statistics.hpp"
-#include "../src/Util.hpp"
 #include "TestUtil.hpp"
+#include "core/Statistics.hpp"
+#include "core/Util.hpp"
 #include "core/fmtmacros.hpp"
 
 #include "third_party/doctest.h"
